+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added collection option to *Execution Triggers*.
 - Added *Sound Falloff* node.
 - Added *Subprograms* to the search menu.
+- Added missing nodes to the node menu
 
 ### Fixed
 
@@ -26,3 +27,4 @@
 - Started following Blender's class naming conventions.
 - Made dependent nodes unsearchable like *Loop Generator* and *Group Output* nodes.
 - Started following Blender's class naming conventions.
+- Changed location of Animation Nodes to Animation Nodes Editor.